+++ CHANGELOG.md
@@ -14,3 +14,4 @@ YouTube music upload support
 
 ## v1.0.2 
 - argparse support
+- Fixed an <a href="https://github.com/sigma67/ytmusicapi/issues/6">upload issue</a> where uploading albums with more than 13 songs caused a response error.